+++ wled_ros_driver/wled_ros_driver/start_wled_service.py
@@ -18,6 +18,7 @@ from wled_ros_driver.types import (
 from wled_ros_driver.config.ros_params import RosParams
 from rcl_interfaces.msg import SetParametersResult
 from dataclasses import asdict
+from time import sleep
 
 
 class AsyncServiceWledNode(Node):
@@ -55,7 +56,7 @@ class AsyncServiceWledNode(Node):
         )
 
         self.get_logger().info("Async service node started")
-        self.get_logger().info("IP address " + self.wled_url)
+        self.get_logger().info(f"IP address {self.wled_url}")
 
     def _load_variables(self):
         """
@@ -72,10 +73,17 @@ class AsyncServiceWledNode(Node):
             .get_parameter_value()
             .integer_value
         )
-        loop = asyncio.get_event_loop()
-        self.sections = loop.run_until_complete(
-            self._load_segments_from_wled_controller(self.wled_url)
-        )
+
+        while rclpy.ok():
+            try:
+                loop = asyncio.get_event_loop()
+                self.sections = loop.run_until_complete(
+                    self._load_data_from_wled_controller()
+                )
+                break
+            except Exception as e:
+                self.get_logger().error(f"{e}. Retrying in 2 seconds...")
+                sleep(2.0)
 
         loaded_scenes = {}
         scenes_params = self.get_parameters_by_prefix(RosParams.SCENES_YAML_NAME)
@@ -93,9 +101,9 @@ class AsyncServiceWledNode(Node):
         for scene_name in loaded_scenes.keys():
             self.scenes[scene_name] = SceneData(**loaded_scenes[scene_name])
 
-    async def _load_segments_from_wled_controller(self) -> dict:
+    async def _load_data_from_wled_controller(self) -> dict:
         """
-        Asynchronous method to fetch segments configuration from WLED controller.
+        Asynchronous method to fetch configuration from WLED controller.
 
         Returns:
             Dict<String,SectionData>
@@ -111,30 +119,40 @@ class AsyncServiceWledNode(Node):
                     segment.segment_id, segment.start, segment.stop
                 )
                 i += 1
+            self.effects = {}
+
+            for effect in device.effects:
+                self.effects[int(effect.effect_id)] = effect.name
             return loaded_sections
 
-    def _parameter_callback(self, params: dict) -> SetParametersResult:
+    def _parameter_callback(self, params: list) -> SetParametersResult:
         """
         Callback function triggered every time ROS parameter is changed while node is running.
         """
         for param in params:
-            self.get_logger().info("param name " + param.name)
-
             if param.name == RosParams.WLED_CONTROLLER_URL:
                 self.wled_url = param.value
-                self.get_logger().info("Updated led ip: " + self.wled_url)
+                self.get_logger().info(f"Updated led ip: {self.wled_url}")
             if param.name == RosParams.WLED_SEGMENTS_COUNT:
                 self.led_count = param.value
-                self.get_logger().info("Updated led number: " + self.wled_url)
+                self.get_logger().info(f"Updated led number: {self.led_count}")
 
             if param.name.startswith(f"{RosParams.SCENES_YAML_NAME}."):
+                self.get_logger().info("Updated scenes")
+
                 values = param.name.split(".")
                 if len(values) == 3:
                     _, scene_name, param_name = values
                     if scene_name not in self.scenes:
                         return SetParametersResult(successful=False)
                     edited_scene = asdict(self.scenes[scene_name])
-                    edited_scene[param_name] = param.value
+                    if param_name == RosParams.SCENES_COLOR_PARAMETER:
+                        edited_scene[param_name] = Color(
+                            param.value[0], param.value[1], param.value[2]
+                        )
+                    else:
+                        edited_scene[param_name] = param.value
+
                     self.scenes[scene_name] = SceneData(**edited_scene)
 
                 else:
@@ -169,10 +187,10 @@ class AsyncServiceWledNode(Node):
         - Turns on the master switch for LEDs.
 
         Args:
-            pars (dict): Dictionary containing following parameters: brightness, start_led_id, stop_led_id, color[red, green, blue].
+            pars (RunLightsData): Object containing following parameters: brightness, start_led_id, stop_led_id, color[red, green, blue].
 
         Returns:
-            str: Confirmation message indicating the scene was set.
+            bool, str: Confirmation message indicating the scene was set.
         """
         self.get_logger().info(f"{pars}")
         try:
@@ -183,13 +201,14 @@ class AsyncServiceWledNode(Node):
                     segment_id=pars.section_id,
                     start=pars.start_led_id,
                     stop=pars.stop_led_id,
-                    color_primary=pars.color,
+                    color_primary=pars.color.as_list,
                     transition=1,
+                    effect=pars.effect,
                 )
                 await led.master(on=True)
             return True, "Scene complete"
         except Exception as e:
-            self.get_logger().error(f"Failed to fetch WLED info: {e}")
+            self.get_logger().error(f"Failed to execute scene_x: {e}")
             return False, "Failed to execute scene"
 
     async def scene_all(self, pars: RunLightsData) -> tuple[bool, str]:
@@ -208,51 +227,66 @@ class AsyncServiceWledNode(Node):
                         segment_id=section.section_id,
                         start=section.start_led_id,
                         stop=section.stop_led_id,
-                        color_primary=pars.color,
+                        color_primary=pars.color.as_list,
                         transition=1,
+                        effect=pars.effect,
                     )
-                    await led.master(on=True)
+                await led.master(on=True)
             return True, "Scene complete"
         except Exception as e:
-            self.get_logger().error(f"Failed to fetch WLED info: {e}")
+            self.get_logger().error(f"Failed to execute scene_all: {e}")
             return False, "Failed to execute scene"
 
     async def scene_off(self, pars: RunLightsData) -> tuple[bool, str]:
         """
-        Asynchronous method to turn off all LEDs using the WLED API.
+        Asynchronous method to turn off selected LEDs using the WLED API.
 
         Args:
-            _: Unused parameter, kept for interface consistency.
+            pars (RunLightsData): Object containing following parameters: brightness, start_led_id, stop_led_id, color[red, green, blue].
 
         Returns:
-            str: Confirmation message indicating the scene is turned off.
+            bool, str: Confirmation message indicating the scene was set.
         """
         try:
             async with WLED(self.wled_url) as led:
                 await led.segment(on=False, segment_id=pars.section_id)
             return True, "Scene 'OFF' complete"
         except Exception as e:
-            self.get_logger().error(f"Failed to fetch WLED info: {e}")
+            self.get_logger().error(f"Failed to execute scene_off: {e}")
             return False, "Failed to execute scene 'OFF'"
 
-    async def scene_off_all(self, pars: RunLightsData) -> tuple[bool, str]:
+    async def scene_off_all(self, _pars: RunLightsData) -> tuple[bool, str]:
         """
         Asynchronous method to turn off all LEDs using the WLED API.
 
         Args:
-            _: Unused parameter, kept for interface consistency.
+            _pars (RunLightsData): Unused parameter kept for interface consistency.
 
         Returns:
-            str: Confirmation message indicating the scene is turned off.
+            bool, str: Confirmation message indicating the scene was set.
         """
         try:
             async with WLED(self.wled_url) as led:
                 for section in self.sections.values():
                     await led.segment(on=False, segment_id=section.section_id)
+                await led.master(on=False)
+
             return True, "Scene 'OFF' complete"
         except Exception as e:
             self.get_logger().error(f"Failed to fetch WLED info: {e}")
             return False, "Failed to execute scene 'OFF'"
+
+    async def do_nothing(self, _pars: RunLightsData) -> tuple[bool, str]:
+        """
+        Empty method to be executed on wrong parameter entry.
+        Args:
+            _pars (RunLightsData): Unused parameter kept for interface consistency.
+
+        Returns:
+            bool, str: Error message indicating that scene key was not recognized.
+
+        """
+        return False, "Failed to recognize section"
 
     def _handle_service(
         self, request: ChangeScene.Request, response: ChangeScene.Response
@@ -296,10 +330,11 @@ class AsyncServiceWledNode(Node):
             SceneFunction.SCENE_OFF: self.scene_off,
             SceneFunction.CHANGE_ALL: self.scene_all,
             SceneFunction.SCENE_OFF_ALL: self.scene_off_all,
+            SceneFunction.NO_CHANGE: self.do_nothing,
         }
 
         self.get_logger().info(
-            f"Requested scene: {request.scene} | section:{request.section} | params: {request.optional_params}"
+            f"Requested scene: {request.scene} | section:{request.section} |  effect_id:{request.effect_id} |params: {request.optional_params}"
         )
 
         params = self._prepare_request_params(request)
@@ -329,8 +364,8 @@ class AsyncServiceWledNode(Node):
             if hasattr(request, RosParams.REQUEST_SECTION_KEY) and request.section
             else RosParams.SECTION_ALL_KEY
         )
-        section_data = {}
-        scene_data = {}
+        section_data = SectionData()
+        scene_data = SceneData()
 
         if scene_key == RosParams.SCENE_CUSTOM_KEY:
             scene_function = SceneFunction.CHANGE_SCENE
@@ -341,29 +376,38 @@ class AsyncServiceWledNode(Node):
         # preset scene
         elif scene_key in self.scenes.keys():
             scene_function = SceneFunction.CHANGE_SCENE
-            scene_data = asdict(self.scenes[scene_key])
+            scene_data = self.scenes[scene_key]
+
         # default behaviour
         else:
             scene_function = SceneFunction.SCENE_OFF
 
         # select section from preset
         if section_key in self.sections.keys():
-            section_data = asdict(self.sections[section_key])
+            section_data = self.sections[section_key]
 
-        # select all sections
-        else:
+        elif section_key == RosParams.SECTION_ALL_KEY:
             if scene_function == SceneFunction.SCENE_OFF:
                 scene_function = SceneFunction.SCENE_OFF_ALL
             else:
                 scene_function = SceneFunction.CHANGE_ALL
 
+        # behaviour when section key not recognized
+        else:
+            scene_function = SceneFunction.NO_CHANGE
+
+        effect = self.effects.get(int(request.effect_id), "Solid")
         return RunLightsData(
-            scene_function=scene_function, **scene_data, **section_data
+            scene_function=scene_function,
+            effect=effect,
+            color=scene_data.color,
+            brightness=scene_data.brightness,
+            **asdict(section_data),
         )
 
-    def _parse_scene_params(self, params_list: list) -> dict:
+    def _parse_scene_params(self, params_list: list) -> SceneData:
         """
-        Parse a list of string parameters to create scene data.
+        Safe parse a list of string parameters to create SceneData object.
 
         Parameters (all optional, default values used if missing or invalid):
             params_list[0]: brightness (int, default 255)
@@ -378,14 +422,22 @@ class AsyncServiceWledNode(Node):
         }
 
         """
-        return {
-            "brightness": int(params_list[0]) if len(params_list) > 0 else 255,
-            "color": Color(
-                R=int(params_list[1]) if len(params_list) > 1 else 255,
-                G=int(params_list[2]) if len(params_list) > 2 else 255,
-                B=int(params_list[3]) if len(params_list) > 3 else 255,
+
+        def get_val(idx):
+            try:
+                return int(params_list.get(idx, 255))
+            except ValueError:
+                self.get_logger().error(f"failed to parse custom data (idx: {idx})")
+                return 255
+
+        return SceneData(
+            brightness=get_val(0),
+            color=Color(
+                R=get_val(1),
+                G=get_val(2),
+                B=get_val(3),
             ),
-        }
+        )
 
 
 def main(args=None):